+++ packages/backend/src/server/web/views/base-embed.tsx
@@ -78,7 +78,7 @@ export function BaseEmbed(props: PropsWithChildren<CommonProps<{
 							Please turn on your JavaScript
 						</p>
 					</noscript>
-					<Splash icon={props.icon} />
+					<Splash icon={props.icon} customSplashText={props.customSplashText} />
 					{props.children}
 				</body>
 			</html>